Allow input payload and output image to be empty
In some cases, like HW key signing done outside op-build, it is not necessary
to specify an input payload (since the hash doesn't matter) or an output
image (since the purpose of this operation is to export the signature). This
patch allow the user to specify __none for --protectedPayload and --out.

Signed-off-by: Dave Heller <hellerda@us.ibm.com>