Close all file descriptors on error exit
When die() is called, close all open fd > 2 prior to exit.  To keep it fast,
only the first sixteen fds are checked, since there would unlikely be more
than that.  A more robust approach would be to check all up to FD_SETSIZE,
as defined in select.h or typesizes.h

Signed-off-by: Dave Heller <hellerda@linux.vnet.ibm.com>