Fix mmap() offset bug and mmap() error detection.
Calls to mmap() with offset are problematic since offset must be on an even
page boundary, which may vary between systems.  Rather than adding code to
handle this, it's easier to not use offset in this case.

Also, the code should use MAP_FAILED instead of NULL, when chekcing mmap() rc.

Also, this patch allows input payload to be omitted, in create-container.c

Signed-off-by: Dave Heller <hellerda@linux.vnet.ibm.com>