Add --validate-ignore-remainder option to print-container
When --validate is requested the default behavior is to use the actual
payload size for calculating the payload hash.  With this option it will
use the payload size from the container header and ignore any additional
bytes following end of payload.  Useful for validataing container files
with padding after the payload.