Support inter-mode import
Supports the case where HW key signing is done standalone in Local or
Independent mode, and the signatures are imported into Production mode.
The use of "signing project" names, used previously in Production only,
is now extended to Local or Independent mode.  The HW key signer may set
a "hw_signing_project_basename".  If set, the export will be prepared
for import to a Production mode process where the same project basename
is in use.  On import, signtool will pick up the artifacts generated
by this basename.