Support inter-mode import (part 2)
This completes support for the the case where HW key signing is done
standalone in Local or Independent mode, and the signatures are imported
into Production mode.  It also adds support for the opposite case, where
HW key siging is done in Production mode and artifacts are imported and
exported to FW key signing in Local mode.  Refactors most of the
keyfinder/sigfinder code in crtSignedContainer.sh.