Support multiple KMS in Production mode
This adds support for PKCS11 as an alternate key management system in
place of signframework, and adds the configuation property SB_KMS and
a new command line option --kms to select between them.  If unset the
default is "signframework".  If set to "pkcs11" two additional
configuration properties, specifying the token name and the shared
library implementing the token, are recognized.  These properties are
set by environment, or via the INI under a new section [pkcs11].