Revert: Change cmdline options --hwPrivKeyX to --hwKeyX (part 2)
This change is dependent on the related change to genPnorImages.pl in
Hostboot, and although it's changed upstream, there are still active
op-build machine configs that are locked to an older version.  So it's
unclear when this change can be finalized.  For now I'm going to punt
and allow the old cmdline options to still work, and continue to
support both the new and deprecated cmdline options.

Signed-off-by: Dave Heller <hellerda@us.ibm.com>