Add support to --sign-project-config for multiple .ini files
In more complex environments, the user, server, and project settings
specified in the .ini file might change depending on the circumstance of
the build.

- The builder might be a different individual and thus have a
different ID to the signing server.
- A different signing server needs to be used

Instead of requiring an .ini file for every possible combination, allow
the .ini files to be broken apart and provided to crtSignedContainer.sh
as separate files.

./crtSignedContainer.sh --sign-project-config \
                        "myuser.ini,myserver.ini,project.ini"