Update crtSignedContainer.sh to ignore SB_PROJECT_INI
For P11 we have the same script doing both dev v1 and production v3 signing.  When
doing production signing the SB_PROJECT_INI env variable is set which includes
the HW key tarball to use for production.  Unfortunately this means the dev v1
signing tries to inject the v3 HW key tarball which doesn't work.

This change ignores the SB_PROJECT_INI when not doing production signing as that
is the only use case for it.