gendilsig: Add support for --pure option for proper ML-DSA pure mode signing
Add support for --pure option to select ML-DSA pure mode signing. In this
case the input data are read from the file are not a digest but the raw
data, which may be of any size. If pure mode is not chosen (the default),
then ensure that the input data read from the file correspond to the
expected size of the SHA3-512 or SHA512 hashes (each 64 bytes).

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>