container: Add support for proper pure mode signing using --pure option
Add a variant of signing with ML-DSA in pure mode where the raw data are
passed to the ML-DSA signing function rather than a digest. This then
achieves proper ML-DSA signing.

ML-DSA in pure mode is supposed to be signing the raw data rather than
hashed data. However, there is also pre-hash (HashML) mode where hashed
data are signed but in this case an object identifier (OID) is part of
the data that are signed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>